TIP-842: Move attribute option normalizer to Akeneo\Pim\Structure
