+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/normalizers.yml
@@ -1,4 +1,5 @@
 services:
+    ### Attribute
     pim_catalog.normalizer.standard.attribute:
         class: 'Akeneo\Pim\Structure\Component\Normalizer\Standard\AttributeNormalizer'
         arguments:
@@ -30,3 +31,29 @@ services:
             - '@pim_versioning.serializer.normalizer.flat.label_translation'
         tags:
             - { name: pim_versioning.serializer.normalizer, priority: 90 }
+
+    ### Attribute option
+    pim_enrich.normalizer.attribute_option:
+        class: 'Akeneo\Pim\Structure\Component\Normalizer\InternalApi\AttributeOptionNormalizer'
+        arguments:
+            - '@pim_catalog.repository.locale'
+            - '@pim_catalog.factory.attribute_option_value'
+        tags:
+            - { name: pim_internal_api_serializer.normalizer }
+
+    pim_api.normalizer.attribute_option:
+        class: 'Akeneo\Pim\Structure\Component\Normalizer\ExternalApi\AttributeOptionNormalizer'
+        arguments:
+            - '@pim_catalog.normalizer.standard.attribute_option'
+        tags:
+            - { name: pim_serializer.normalizer, priority: 90 }
+
+    pim_versioning.serializer.normalizer.flat.option:
+        class: 'Akeneo\Pim\Structure\Component\Normalizer\Versionning\AttributeOptionNormalizer'
+        tags:
+            - { name: pim_versioning.serializer.normalizer, priority: 90 }
+
+    pim_catalog.normalizer.standard.attribute_option:
+        class: 'Akeneo\Pim\Structure\Component\Normalizer\Standard\AttributeOptionNormalizer'
+        tags:
+            - { name: pim_serializer.normalizer, priority: 90 }
