+++ src/Akeneo/Pim/Structure/Component/spec/Normalizer/Standard/AttributeOptionNormalizerSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Component\Catalog\Normalizer\Standard;
+namespace spec\Akeneo\Pim\Structure\Component\Normalizer\Standard;
 
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
@@ -11,7 +11,7 @@ class AttributeOptionNormalizerSpec extends ObjectBehavior
 {
     function it_is_initializable()
     {
-        $this->shouldHaveType('Pim\Component\Catalog\Normalizer\Standard\AttributeOptionNormalizer');
+        $this->shouldHaveType('Akeneo\Pim\Structure\Component\Normalizer\Standard\AttributeOptionNormalizer');
     }
 
     function it_is_a_normalizer()
