+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/normalizers.yml
@@ -6,7 +6,6 @@ parameters:
     pim_api.normalizer.family.class: Akeneo\Tool\Component\Api\Normalizer\FamilyNormalizer
     pim_api.normalizer.family_variant.class: Akeneo\Tool\Component\Api\Normalizer\FamilyVariantNormalizer
     pim_api.normalizer.category.class: Akeneo\Tool\Component\Api\Normalizer\CategoryNormalizer
-    pim_api.normalizer.attribute_option.class: Akeneo\Tool\Component\Api\Normalizer\AttributeOptionNormalizer
     pim_api.normalizer.attribute_group.class: Akeneo\Tool\Component\Api\Normalizer\AttributeGroupNormalizer
     pim_api.normalizer.file.class: Akeneo\Tool\Component\Api\Normalizer\FileNormalizer
     pim_api.normalizer.association_type.class: Akeneo\Tool\Component\Api\Normalizer\AssociationTypeNormalizer
@@ -67,13 +66,6 @@ services:
         class: '%pim_api.normalizer.category.class%'
         arguments:
             - '@pim_catalog.normalizer.standard.category'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_api.normalizer.attribute_option:
-        class: '%pim_api.normalizer.attribute_option.class%'
-        arguments:
-            - '@pim_catalog.normalizer.standard.attribute_option'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
