+++ src/Pim/Bundle/CatalogBundle/Resources/config/serializers_standard.yml
@@ -11,7 +11,6 @@ parameters:
     pim_catalog.normalizer.standard.file.class: Pim\Component\Catalog\Normalizer\Standard\FileNormalizer
     pim_catalog.normalizer.standard.association_type.class: Pim\Component\Catalog\Normalizer\Standard\AssociationTypeNormalizer
     pim_catalog.normalizer.standard.attribute_group.class: Pim\Component\Catalog\Normalizer\Standard\AttributeGroupNormalizer
-    pim_catalog.normalizer.standard.attribute_option.class: Pim\Component\Catalog\Normalizer\Standard\AttributeOptionNormalizer
     pim_catalog.normalizer.standard.category.class: Pim\Component\Catalog\Normalizer\Standard\CategoryNormalizer
     pim_catalog.normalizer.standard.family.class: Pim\Component\Catalog\Normalizer\Standard\FamilyNormalizer
     pim_catalog.normalizer.standard.family_variant.class: Pim\Component\Catalog\Normalizer\Standard\FamilyVariantNormalizer
@@ -96,11 +95,6 @@ services:
         arguments:
             - '@pim_catalog.normalizer.standard.translation'
             - '@pim_catalog.repository.attribute'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_catalog.normalizer.standard.attribute_option:
-        class: '%pim_catalog.normalizer.standard.attribute_option.class%'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
