+++ src/Pim/Bundle/EnrichBundle/Resources/config/normalizers.yml
@@ -1,6 +1,5 @@
 parameters:
     pim_enrich.normalizer.attribute_option_value_collection.class: Pim\Bundle\EnrichBundle\Normalizer\AttributeOptionValueCollectionNormalizer
-    pim_enrich.normalizer.attribute_option.class:                  Pim\Bundle\EnrichBundle\Normalizer\AttributeOptionNormalizer
     pim_enrich.normalizer.attribute_option_value.class:            Pim\Bundle\EnrichBundle\Normalizer\AttributeOptionValueNormalizer
     pim_enrich.normalizer.product.class:                           Pim\Bundle\EnrichBundle\Normalizer\ProductNormalizer
     pim_enrich.normalizer.product_model.class:                     Pim\Bundle\EnrichBundle\Normalizer\ProductModelNormalizer
@@ -30,14 +29,6 @@ parameters:
 services:
     pim_enrich.normalizer.attribute_option_value_collection:
         class: '%pim_enrich.normalizer.attribute_option_value_collection.class%'
-        tags:
-            - { name: pim_internal_api_serializer.normalizer }
-
-    pim_enrich.normalizer.attribute_option:
-        class: '%pim_enrich.normalizer.attribute_option.class%'
-        arguments:
-            - '@pim_catalog.repository.locale'
-            - '@pim_catalog.factory.attribute_option_value'
         tags:
             - { name: pim_internal_api_serializer.normalizer }
